requirements.txt
streamlit
openai
requests
pillow
PyPDF2
gTTS
SpeechRecognition
streamlit-webrtc
python-dotenv
